python: move VMIHandler into it's own package
