+++ python/microvmi/volatility/__init__.py
@@ -1,0 +1,8 @@
+"""
+This package regroups the volatility plugins developed on top of the Microvmi package
+
+It is intended to be used as a volatility3 plugins directory:
+
+Example:
+    vol.py -p /path/to/microvmi/volatility windows.pslist.PsList
+"""
