+++ python/setup.cfg
@@ -25,5 +25,5 @@ source = microvmi
 [mypy-microvmi.pymicrovmi.*]
 ignore_missing_imports = True
 
-[mypy-volatility3.framework.layers.handler.*]
+[mypy-volatility3.framework.layers.resources.*]
 ignore_missing_imports = True
